refactore: change name from controller I was planning to create a separede files for all controllers but i decide to join they all
