+++ src/http/http.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './controllers/app.controller';
 import { AppService } from './services/app.service';
-import { CreateTournamentController } from './controllers/create-tournament.controller';
+import { CreateTournamentController } from './controllers/tournaments.controller';
 import { CreateTournamentUseCase } from 'src/core/use-cases/create-tournament-use-case';
 
 @Module({
